check_whence.py: Add function to get listing of files
Commit [1] introduces check_whence.py inside
copy-firmware.sh before the firmware is consumed.

But current check_whence.py supports only git
repo, and it will cause error when linux-firmware tar
is used, as it will not be a git repo.

So adding a check to parse local directory to create
a set of files to validate it with WHENCE file.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git/commit/?id=6edd0fd36acb93c8075bbd700ad91f53fc810c0e

Signed-off-by: Vivek Puar <quic_vpuar@quicinc.com>